refactor: reorganize tests
